Move the CLI update check into an updatecheck package
Extract the update check, and the command metadata it sends with the version request, out of `pkg/cmd/pulumi/pulumi.go` into a new `pkg/cmd/pulumi/updatecheck` package behind a small `Start`/`Finish` API, and move its tests along with it. No functional changes.
