+++ pkg/cmd/pulumi/pulumi.go
@@ -15,31 +15,19 @@
 package main
 
 import (
-	"bytes"
-	"context"
 	"encoding/csv"
-	"encoding/json"
-	"errors"
 	"fmt"
 	"io"
 	"log"
 	"log/slog"
-	"net/http"
-	"net/url"
 	"os"
-	"os/exec"
-	"os/user"
-	"path/filepath"
-	"regexp"
 	"runtime"
-	"strconv"
 	"strings"
 	"time"
 
 	"github.com/opentracing/opentracing-go"
 
 	"github.com/blang/semver"
-	"github.com/djherbis/times"
 	"github.com/moby/term"
 
 	"github.com/spf13/cobra"
@@ -81,22 +69,19 @@ import (
 	"github.com/pulumi/pulumi/pkg/v3/cmd/pulumi/state"
 	"github.com/pulumi/pulumi/pkg/v3/cmd/pulumi/templatecmd"
 	"github.com/pulumi/pulumi/pkg/v3/cmd/pulumi/trace"
+	"github.com/pulumi/pulumi/pkg/v3/cmd/pulumi/updatecheck"
 	cmdVersion "github.com/pulumi/pulumi/pkg/v3/cmd/pulumi/version"
 	"github.com/pulumi/pulumi/pkg/v3/cmd/pulumi/whoami"
 	backendlogging "github.com/pulumi/pulumi/pkg/v3/logging"
 	"github.com/pulumi/pulumi/pkg/v3/util/tracing"
 	pkgWorkspace "github.com/pulumi/pulumi/pkg/v3/workspace"
-	"github.com/pulumi/pulumi/sdk/v3/go/common/diag"
-	"github.com/pulumi/pulumi/sdk/v3/go/common/diag/colors"
 	"github.com/pulumi/pulumi/sdk/v3/go/common/env"
 	"github.com/pulumi/pulumi/sdk/v3/go/common/util/agentdetect"
 	"github.com/pulumi/pulumi/sdk/v3/go/common/util/cmdutil"
 	"github.com/pulumi/pulumi/sdk/v3/go/common/util/contract"
 	declared "github.com/pulumi/pulumi/sdk/v3/go/common/util/env"
-	"github.com/pulumi/pulumi/sdk/v3/go/common/util/httputil"
 	"github.com/pulumi/pulumi/sdk/v3/go/common/util/logging"
 	"github.com/pulumi/pulumi/sdk/v3/go/common/version"
-	"github.com/pulumi/pulumi/sdk/v3/go/common/workspace"
 	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/propagation"
@@ -220,7 +205,7 @@ func NewPulumiCmd() (*cobra.Command, func()) {
 
 	processStartTime := time.Now()
 
-	updateCheckResult := make(chan *updateCheckResult)
+	var updateCheck <-chan *updatecheck.CheckResult
 
 	var cmd *cobra.Command
 	cleanup := func() {
@@ -280,15 +265,6 @@ func NewPulumiCmd() (*cobra.Command, func()) {
 			"\n" +
 			"For more information, please visit the project page: https://www.pulumi.com/docs/",
 		PersistentPreRunE: func(cmd *cobra.Command, args []string) error {
-			// If we fail before we start the async update check, go ahead and close the
-			// channel since we know it will never receive a value.
-			var waitForUpdateCheck bool
-			defer func() {
-				if !waitForUpdateCheck {
-					close(updateCheckResult)
-				}
-			}()
-
 			// Commands like `pulumi do` set DisableFlagParsing on themselves so they can
 			// build a dynamic flag tree from a provider's schema. cobra skips flag parsing
 			// entirely for such commands, which means our root persistent flag variables
@@ -357,7 +333,7 @@ func NewPulumiCmd() (*cobra.Command, func()) {
 				ctx = tracing.ContextWithOptions(ctx, tracingOptions)
 			}
 
-			metadata := getCLIMetadata(cmd, os.Environ(), args)
+			metadata := updatecheck.GetCLIMetadata(cmd, os.Environ(), args)
 			slog.InfoContext(ctx, "CLI Metadata", "metadata", metadata)
 
 			if cmdutil.IsOTelEnabled() {
@@ -408,17 +384,10 @@ func NewPulumiCmd() (*cobra.Command, func()) {
 				}
 			}
 
-			if env.SkipUpdateCheck.Value() {
-				slog.InfoContext(ctx, "skipping update check")
-			} else {
-				// Run the version check in parallel so that it doesn't block executing the command.
-				// If there is a new version to report, we will do so after the command has finished.
-				waitForUpdateCheck = true
-				go func() {
-					updateCheckResult <- checkForUpdate(ctx, client.PulumiCloudURL, metadata)
-					close(updateCheckResult)
-				}()
-			}
+			// The version check runs in the background so that it doesn't block executing the
+			// command. If there is a new version to report, we will do so after the command has
+			// finished.
+			updateCheck = updatecheck.Start(ctx, client.PulumiCloudURL, metadata)
 
 			return nil
 		},
@@ -430,17 +399,8 @@ func NewPulumiCmd() (*cobra.Command, func()) {
 			isJSON := jsonFlag != nil && jsonFlag.Value.String() == "true"
 
 			if !isJSON {
-				select {
-				case result, ok := <-updateCheckResult:
-					if ok && result != nil {
-						cmdutil.Diag().Warningf(result.diag)
-						err := cacheVersionInfo(result.versionInfo)
-						if err != nil {
-							slog.Info("failed to cache version info", "err", err)
-						}
-					}
-				default:
-					// The version didn't make it in time, so don't block the user
+				if result := updatecheck.Finish(updateCheck); result != nil {
+					cmdutil.Diag().Warningf(result.Diag)
 				}
 			}
 		},
@@ -605,497 +565,6 @@ func NewPulumiCmd() (*cobra.Command, func()) {
 	rattler.Install(cmd)
 
 	return cmd, cleanup
-}
-
-// haveNewerDevVersion checks whether we have a newer dev version available.
-func haveNewerDevVersion(devVersion semver.Version, curVersion semver.Version) bool {
-	if devVersion.Major != curVersion.Major {
-		return devVersion.Major > curVersion.Major
-	}
-	if devVersion.Minor != curVersion.Minor {
-		return devVersion.Minor > curVersion.Minor
-	}
-	if devVersion.Patch != curVersion.Patch {
-		return devVersion.Patch > curVersion.Patch
-	}
-
-	// The dev version string looks like: v1.0.0-11-g4ff08363.  We
-	// can determine whether we have a newer dev version by
-	// comparing the second part of the version string, which is
-	// the number of commits since the last tag.
-	devVersionParts := strings.Split(devVersion.String(), "-")
-	curVersionParts := strings.Split(curVersion.String(), "-")
-
-	// We're being leninent with parsing here.  If we can't parse
-	// a version number correctly for any reason, we default to
-	// pretending there is no newer version, and not warning the
-	// user.  As this is only a warning this is better than
-	// asserting or crashing in the error case.
-	if len(devVersionParts) != 3 || len(curVersionParts) != 3 {
-		return false
-	}
-	devCommits, err := strconv.Atoi(devVersionParts[1])
-	if err != nil {
-		return false
-	}
-	curCommits, err := strconv.Atoi(curVersionParts[1])
-	if err != nil {
-		return false
-	}
-	return devCommits > curCommits
-}
-
-type updateCheckResult struct {
-	diag        *diag.Diag
-	versionInfo cachedVersionInfo
-}
-
-// checkForUpdate checks to see if the CLI needs to be updated, and if so emits a warning, as well as information
-// as to how it can be upgraded.
-func checkForUpdate(ctx context.Context, cloudURL string, metadata map[string]string) *updateCheckResult {
-	curVer, err := semver.ParseTolerant(version.Version)
-	if err != nil {
-		slog.InfoContext(ctx, "error parsing current version", "err", err)
-	}
-
-	// We don't care about warning about updates if this is a locally-compiled version
-	if isLocalVersion(curVer) {
-		return nil
-	}
-
-	isCurVerDev := isDevVersion(curVer)
-	canPrompt, lastPromptTimestampMS := checkVersionPrompt(isCurVerDev)
-
-	latestVer, oldestAllowedVer, devVer, err := getCLIVersionInfo(ctx, cloudURL, metadata)
-	if err != nil {
-		slog.InfoContext(ctx, fmt.Sprintf("error fetching latest version information; set %s to true to skip update checks",
-			env.SkipUpdateCheck.Var().Name()), "err", err)
-	}
-
-	willPrompt := canPrompt &&
-		((isCurVerDev && haveNewerDevVersion(devVer, curVer)) ||
-			(!isCurVerDev && oldestAllowedVer.GT(curVer)))
-
-	if willPrompt {
-		lastPromptTimestampMS = time.Now().UnixMilli() // We're prompting, update the timestamp
-	}
-
-	if willPrompt {
-		if isCurVerDev {
-			latestVer = devVer
-		}
-
-		msg := getUpgradeMessage(latestVer, curVer, isCurVerDev)
-		return &updateCheckResult{
-			diag: diag.RawMessage("", msg),
-			versionInfo: cachedVersionInfo{
-				LatestVersion:         latestVer.String(),
-				OldestWithoutWarning:  oldestAllowedVer.String(),
-				LatestDevVersion:      devVer.String(),
-				LastPromptTimeStampMS: lastPromptTimestampMS,
-			},
-		}
-	}
-
-	return nil
-}
-
-// getCLIMetadata returns a map of metadata about the given CLI command.
-func getCLIMetadata(cmd *cobra.Command, environ []string, args []string) map[string]string {
-	if cmd == nil {
-		return nil
-	}
-
-	command := cmd.CommandPath()
-
-	var flags strings.Builder
-	i := 0
-	cmd.Flags().VisitAll(func(f *pflag.Flag) {
-		if f.Changed {
-			if i > 0 {
-				flags.WriteRune(' ')
-			}
-			flags.WriteString("--" + f.Name)
-			i++
-		}
-	})
-
-	if command == "pulumi plugin run" {
-		if len(args) > 0 {
-			positionals := positionalArgs(cmd, args)
-			if len(positionals) > 0 {
-				command += " " + positionals[0]
-			}
-		}
-	}
-
-	if command == "pulumi do" {
-		positionals := positionalArgs(cmd, args)
-		// Include the resource/function token
-		if len(positionals) > 0 {
-			command += " " + positionals[0]
-		}
-		if len(positionals) > 1 {
-			switch positionals[1] {
-			case "create", "read", "patch", "delete", "list":
-				command += " " + positionals[1]
-			}
-		}
-	}
-
-	env := []string{}
-	for _, e := range environ {
-		parts := strings.Split(e, "=")
-		if len(parts) == 2 && strings.HasPrefix(parts[0], "PULUMI_") {
-			env = append(env, parts[0])
-		}
-	}
-	envString := strings.Join(env, " ")
-
-	metadata := map[string]string{
-		"Command":     command,
-		"Flags":       flags.String(),
-		"Environment": envString,
-	}
-
-	return metadata
-}
-
-// positionalArgs extracts positional arguments from a raw argument list,
-// filtering out flags. This is used for commands with DisableFlagParsing
-// where args contains everything including flags. Known flags from the
-// command are registered so pflag can skip over their values correctly;
-// unknown flags are also tolerated.
-func positionalArgs(cmd *cobra.Command, args []string) []string {
-	fs := pflag.NewFlagSet("telemetry", pflag.ContinueOnError)
-	fs.ParseErrorsAllowlist.UnknownFlags = true
-	// Register the command's own flags so pflag can correctly skip their
-	// values (e.g. --package "foo" should not treat "foo" as positional).
-	// Visit both Flags() and PersistentFlags() to cover all defined flags;
-	// cobra merges them during Execute(), but we may run before that merge.
-	cmd.Flags().VisitAll(func(f *pflag.Flag) {
-		fs.AddFlag(f)
-	})
-	cmd.PersistentFlags().VisitAll(func(f *pflag.Flag) {
-		if fs.Lookup(f.Name) == nil {
-			fs.AddFlag(f)
-		}
-	})
-	// Silence any output from parsing errors.
-	fs.SetOutput(io.Discard)
-	_ = fs.Parse(args)
-	return fs.Args()
-}
-
-// getCLIVersionInfo returns information about the latest version of the CLI and the oldest version that should be
-// allowed without warning, as well as the amount of time to cache this information.
-func getCLIVersionInfo(
-	ctx context.Context,
-	cloudURL string,
-	metadata map[string]string,
-) (semver.Version, semver.Version, semver.Version, error) {
-	creds, err := workspace.GetStoredCredentials()
-	apiToken := creds.AccessTokens[creds.Current]
-
-	if err != nil || creds.Current != cloudURL {
-		apiToken = ""
-		metadata = nil
-	}
-
-	client := client.NewClient(cloudURL, apiToken, false, cmdutil.Diag())
-	ctx, cancel := context.WithTimeout(ctx, 3*time.Second)
-	defer cancel()
-	latest, oldest, dev, err := client.GetCLIVersionInfo(ctx, metadata)
-	if err != nil {
-		return semver.Version{}, semver.Version{}, semver.Version{}, err
-	}
-
-	brewLatest, isBrew, err := getLatestBrewFormulaVersion()
-	if err != nil {
-		slog.InfoContext(ctx, "error determining if the running executable was installed with brew", "err", err)
-	}
-	if isBrew {
-		// When consulting Homebrew for version info, we just use the latest version as the oldest allowed.
-		latest, oldest, dev = brewLatest, brewLatest, brewLatest
-	}
-
-	// Don't return the err from getLatestBrewFormulaVersion here, we just log that above.
-	return latest, oldest, dev, nil
-}
-
-// cacheVersionInfo saves version information in a cache file to be looked up later.
-func cacheVersionInfo(info cachedVersionInfo) error {
-	updateCheckFile, err := pkgWorkspace.GetCachedVersionFilePath()
-	if err != nil {
-		return err
-	}
-
-	file, err := os.OpenFile(updateCheckFile, os.O_CREATE|os.O_TRUNC|os.O_RDWR, 0o600)
-	if err != nil {
-		return err
-	}
-	defer contract.IgnoreClose(file)
-
-	return json.NewEncoder(file).Encode(info)
-}
-
-// readVersionInfo reads version information from the cache file.
-func readVersionInfo() (cachedVersionInfo, error) {
-	updateCheckFile, err := pkgWorkspace.GetCachedVersionFilePath()
-	if err != nil {
-		return cachedVersionInfo{}, err
-	}
-
-	file, err := os.Open(updateCheckFile)
-	if err != nil {
-		return cachedVersionInfo{}, err
-	}
-	defer contract.IgnoreClose(file)
-
-	var info cachedVersionInfo
-	if err := json.NewDecoder(file).Decode(&info); err != nil {
-		return cachedVersionInfo{}, err
-	}
-
-	return info, nil
-}
-
-// checkVersionPrompt determines if
-//   - enough time has passed since we last prompted the user
-//   - the timestamp when we last prompted the user
-//
-// If we can't read the cached versions file, we return true and a zero time,
-// indicating that we want to possibly prompt the user for an upgrade.
-func checkVersionPrompt(devVersion bool) (bool, int64) {
-	updateCheckFile, err := pkgWorkspace.GetCachedVersionFilePath()
-	if err != nil {
-		return true, 0
-	}
-
-	ts, err := times.Stat(updateCheckFile)
-	if err != nil {
-		return true, 0
-	}
-
-	info, err := readVersionInfo()
-	if err != nil {
-		return true, 0
-	}
-
-	// Prompt at most once a day for regular versions, and at most once an hour for dev versions.
-	promptCacheTime := 24 * time.Hour
-	if devVersion {
-		promptCacheTime = 1 * time.Hour
-	}
-
-	// Fallback to the file modification date if we didn't save a last prompt timestamp yet.
-	lastPrompt := ts.ModTime()
-	if info.LastPromptTimeStampMS > 0 {
-		lastPrompt = time.UnixMilli(info.LastPromptTimeStampMS)
-	}
-
-	nextPrompt := lastPrompt.Add(promptCacheTime)
-	expired := nextPrompt.Before(time.Now())
-
-	return expired, lastPrompt.UnixMilli()
-}
-
-// cachedVersionInfo is the on disk format of the version information the CLI caches between runs.
-type cachedVersionInfo struct {
-	LatestVersion         string `json:"latestVersion"`
-	OldestWithoutWarning  string `json:"oldestWithoutWarning"`
-	LatestDevVersion      string `json:"latestDevVersion"`
-	LastPromptTimeStampMS int64  `json:"LastPromptMS,omitempty"`
-}
-
-// getUpgradeMessage gets a message to display to a user instructing them they are out of date and how to move from
-// current to latest.
-func getUpgradeMessage(latest semver.Version, current semver.Version, isDevVersion bool) string {
-	cmd := getUpgradeCommand(isDevVersion)
-
-	// If the current version is "very old", we'll return a more urgent message. "Very old" is defined as more than 24
-	// minor versions behind when the major versions are the same. Assuming a release cadence of on average 1 minor
-	// version per week, this translates to roughly 6 months. Note that we don't consider major version differences, since
-	// it's hard to know what we'd want to do in those cases. E.g. it might be that a new version of Pulumi is radically
-	// different, rather than "just improved", and so we don't want to warn about that.
-	prefix := "A new version of Pulumi is available."
-
-	minorDiff := diffMinorVersions(current, latest)
-	if minorDiff > 24 {
-		prefix = colors.SpecAttention +
-			"You are running a very old version of Pulumi and should upgrade as soon as possible." + colors.Reset
-	}
-
-	msg := fmt.Sprintf("%s To upgrade from version '%s' to '%s', ", prefix, current, latest)
-	if cmd != "" {
-		msg += "run \n   " + cmd + "\nor "
-	}
-
-	msg += "visit https://pulumi.com/docs/install/ for manual instructions and release notes."
-	return msg
-}
-
-// diffMinorVersions compares two semver versions.
-//   - If the major versions of the two versions are the same, it returns the difference in their minor versions. This
-//     difference will be a positive number if v2 is greater than v1 and a negative number if v1 is greater than v2.
-//   - If the major versions differ, it returns 0.
-func diffMinorVersions(v1 semver.Version, v2 semver.Version) int64 {
-	if v1.Major != v2.Major {
-		return 0
-	}
-
-	return int64(v2.Minor - v1.Minor)
-}
-
-// getUpgradeCommand returns a command that will upgrade the CLI to the newest version. If we can not determine how
-// the CLI was installed, the empty string is returned.
-func getUpgradeCommand(isDevVersion bool) string {
-	curUser, err := user.Current()
-	if err != nil {
-		return ""
-	}
-
-	exe, err := os.Executable()
-	if err != nil {
-		return ""
-	}
-
-	isBrew, err := isBrewInstall(exe)
-	if err != nil {
-		slog.Info("error determining if the running executable was installed with brew", "err", err)
-	}
-	if isBrew {
-		return "$ brew update && brew upgrade pulumi"
-	}
-
-	if filepath.Dir(exe) != filepath.Join(curUser.HomeDir, workspace.BookkeepingDir, "bin") {
-		return ""
-	}
-
-	if runtime.GOOS != "windows" {
-		command := "$ curl -sSL https://get.pulumi.com | sh"
-		if isDevVersion {
-			command = command + " -s -- --version dev"
-		}
-		return command
-	}
-
-	powershellCmd := `"%SystemRoot%\System32\WindowsPowerShell\v1.0\powershell.exe"`
-
-	if _, err := exec.LookPath("powershell"); err == nil {
-		powershellCmd = "powershell"
-	}
-
-	powershellCmd = "> " + powershellCmd + ` -NoProfile -InputFormat None -ExecutionPolicy Bypass -Command "iex ` +
-		`((New-Object System.Net.WebClient).DownloadString('https://get.pulumi.com/install.ps1'))"`
-	if isDevVersion {
-		powershellCmd = powershellCmd + " -version dev"
-	}
-	return powershellCmd
-}
-
-// isBrewInstall returns true if the current running executable is running on macOS and was installed with brew.
-func isBrewInstall(exe string) (bool, error) {
-	if runtime.GOOS != "darwin" {
-		return false, nil
-	}
-
-	exePath, err := filepath.EvalSymlinks(exe)
-	if err != nil {
-		return false, err
-	}
-
-	brewBin, err := exec.LookPath("brew")
-	if err != nil {
-		return false, err
-	}
-
-	brewPrefixCmd := exec.Command(brewBin, "--prefix", "pulumi")
-	var stdout bytes.Buffer
-	var stderr bytes.Buffer
-	brewPrefixCmd.Stdout = &stdout
-	brewPrefixCmd.Stderr = &stderr
-	if err = brewPrefixCmd.Run(); err != nil {
-		if ee, ok := err.(*exec.ExitError); ok {
-			ee.Stderr = stderr.Bytes()
-		}
-		return false, fmt.Errorf("'brew --prefix pulumi' failed: %w", err)
-	}
-
-	brewPrefixCmdOutput := strings.TrimSpace(stdout.String())
-	if brewPrefixCmdOutput == "" {
-		return false, errors.New("trimmed output from 'brew --prefix pulumi' is empty")
-	}
-
-	brewPrefixPath, err := filepath.EvalSymlinks(brewPrefixCmdOutput)
-	if err != nil {
-		return false, err
-	}
-
-	brewPrefixExePath := filepath.Join(brewPrefixPath, "bin", "pulumi")
-	return exePath == brewPrefixExePath, nil
-}
-
-func getLatestBrewFormulaVersion() (semver.Version, bool, error) {
-	exe, err := os.Executable()
-	if err != nil {
-		return semver.Version{}, false, err
-	}
-
-	isBrew, err := isBrewInstall(exe)
-	if err != nil {
-		return semver.Version{}, false, err
-	}
-	if !isBrew {
-		return semver.Version{}, false, nil
-	}
-
-	const formulaJSON = "https://formulae.brew.sh/api/formula/pulumi.json"
-	url, err := url.Parse(formulaJSON)
-	contract.AssertNoErrorf(err, "Could not parse URL %q", formulaJSON)
-
-	resp, err := httputil.DoWithRetry(&http.Request{
-		Method: http.MethodGet,
-		URL:    url,
-	}, http.DefaultClient)
-	if err != nil {
-		return semver.Version{}, false, err
-	}
-	defer contract.IgnoreClose(resp.Body)
-
-	type versions struct {
-		Stable string `json:"stable"`
-	}
-	var formula struct {
-		Versions versions `json:"versions"`
-	}
-	if err := json.NewDecoder(resp.Body).Decode(&formula); err != nil {
-		return semver.Version{}, false, err
-	}
-
-	stable, err := semver.ParseTolerant(formula.Versions.Stable)
-	if err != nil {
-		return semver.Version{}, false, err
-	}
-	return stable, true, nil
-}
-
-func isLocalVersion(s semver.Version) bool {
-	if len(s.Pre) == 0 {
-		return false
-	}
-
-	devStrings := regexp.MustCompile(`alpha|beta|dev|rc`)
-	return !s.Pre[0].IsNum && devStrings.MatchString(s.Pre[0].VersionStr)
-}
-
-func isDevVersion(s semver.Version) bool {
-	if len(s.Pre) == 0 {
-		return false
-	}
-
-	devRegex := regexp.MustCompile(`\d*-g[0-9a-f]*$`)
-	return !s.Pre[0].IsNum && devRegex.MatchString(s.Pre[0].VersionStr)
 }
 
 // We want to expose all the flags on all the commands as configurable with
